pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trainforge-conductor"
version = "0.1.0"
description = "Multi-Provider LLM API Conductor with intelligent rate limit scheduling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "TrainForge"}
]
keywords = ["llm", "api", "conductor", "cerebras", "nvidia", "rate-limiting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.4.0",
    "openai>=1.58.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "requests>=2.32.0",
]

[project.scripts]
trainforge-conductor = "app.main:main"

[project.urls]
Homepage = "https://github.com/trainforge/conductor"
Repository = "https://github.com/trainforge/conductor"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

